requirements.txt
python-telegram-bot
logging
watson-developer-cloud
pyexecjs
Flask
numpy
scipy
sklearn
pandas
tweepy
cloudant
